Avoid double destroy() on widgets like Menu that are logically children of other widgets even though they are attached to <body> and thus picked up as "top level widgets".   Fixes #9966 !strict.
git-svn-id: http://svn.dojotoolkit.org/src/dijit/trunk@20293 560b804f-0ae3-0310-86f3-f6aa0a117693

+++ PopupMenuItem.js
@@ -50,7 +50,11 @@ dojo.declare("dijit.PopupMenuItem",
 		
 		destroyDescendants: function(){
 			if(this.popup){
-				this.popup.destroyRecursive();
+				// Destroy the popup, unless it's already been destroyed.  This can happen because
+				// the popup is a direct child of <body> even though it's logically my child.
+				if(!this.popup._destroyed){
+					this.popup.destroyRecursive();
+				}
 				delete this.popup;
 			}
 			this.inherited(arguments);
