+++ _HasDropDown.js
@@ -201,7 +201,11 @@ dojo.declare("dijit._HasDropDown",
 
 		destroyDescendants: function(){
 			if(this.dropDown){
-				this.dropDown.destroyRecursive();
+				// Destroy the drop down, unless it's already been destroyed.  This can happen because
+				// the drop down is a direct child of <body> even though it's logically my child.
+				if(!this.dropDown._destroyed){
+					this.dropDown.destroyRecursive();
+				}
 				delete this.dropDown;
 			}
 			this.inherited(arguments);
